--- lambda/echoServer/tsconfig.json
@@ -1,17 +1,0 @@
-{
-    "compilerOptions": {
-      "target": "es2020",
-      "strict": true,
-      "preserveConstEnums": true,
-      "noEmit": true,
-      "sourceMap": false,
-      "module":"commonjs",
-      "moduleResolution":"node",
-      "esModuleInterop": true, 
-      "skipLibCheck": true,
-      "forceConsistentCasingInFileNames": true, 
-      "isolatedModules": true, 
-      "baseUrl": "./src"
-    },
-    "exclude": ["node_modules", "**/*.test.ts"]
-  }
